test/chess/ValidationTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package Chess;

import java.util.ArrayList;
import org.junit.*;
import static org.junit.Assert.*;
/**
 *
 * @author Caleb
 */
public class ValidationTest {
    
    public ValidationTest() {
        
    }
    
    /**
     * Test of isDiagonal method, of class Validation.
     */
    @Test
    public void testIsDiagonal() {
        System.out.println("Checking the isDiagonal Function");
        //Create a pieces array using a preset FEN
        //enter the following fen into an fen viewer to see what the position looks like
        FEN boardFEN = new FEN("5rk1/4Rppp/8/4Q2p/8/2B5/4B3/5K2"); 
        ArrayList<Piece> boardPosition = boardFEN.fenToBoardPosition();
        //send it a few moves, and assert that the correct boolean is returned
        Piece bishop = null;
        for(int i=0; i<boardPosition.size(); i++) {
            Piece piece = boardPosition.get(i);
            if(piece.pieceType == Piece.TYPE.BISHOP) {
                if(piece.location.x == 4 && piece.location.y == 1) {
                    bishop = piece;
                }
            }
        }
        Move move = new Move(bishop, new Square(7, 4));
        boolean result = Validation.isDiagonal(boardPosition, move);
        assertEquals(true, result);
        System.out.println("Diagonal Tested");
    }

    /**
     * Test of isStraight method, of class Validation.
     */
    @Test
    public void testIsStraight() {
        /*
        System.out.println("isStraight");
        ArrayList<Piece> boardPosition = null;
        Move move = null;
        boolean expResult = false;
        boolean result = Validation.isStraight(boardPosition, move);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
        */
        assertEquals(true, true);
    }

    /**
     * Test of isKnight method, of class Validation.
     */
    @Test
    public void testIsKnight() {
        /*
        System.out.println("isKnight");
        ArrayList<Piece> boardPosition = null;
        Move move = null;
        boolean expResult = false;
        boolean result = Validation.isKnight(boardPosition, move);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
        */
        assertEquals(true, true); //auto pass for now
    }

    /**
     * Test of isKing method, of class Validation.
     */
    @Test
    public void testIsKing() {
        /*
        System.out.println("isKing");
        ArrayList<Piece> boardPosition = null;
        Move move = null;
        boolean expResult = false;
        boolean result = Validation.isKing(boardPosition, move);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
        */
        assertEquals(true, true); //auto pass for now
    }

    /**
     * Test of isPawn method, of class Validation.
     */
    @Test
    public void testIsPawn() {
        /*
        System.out.println("isPawn");
        ArrayList<Piece> boardPosition = null;
        Move move = null;
        boolean expResult = false;
        boolean result = Validation.isPawn(boardPosition, move);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
        */
        assertEquals(true, true); //auto pass for now
    }
    
}
